# [MERGE] sale: Convert onchanges into compute + improve perfs