--- a/addons/account/wizard/accrued_orders.py
+++ b/addons/account/wizard/accrued_orders.py
@@ -155,7 +155,7 @@ class AccruedExpenseRevenue(models.TransientModel):
                     o.order_line.with_context(accrual_entry_date=self.date)._compute_qty_delivered()
                     o.order_line.with_context(accrual_entry_date=self.date)._compute_qty_invoiced()
                     o.order_line.with_context(accrual_entry_date=self.date)._compute_untaxed_amount_invoiced()
-                    o.order_line.with_context(accrual_entry_date=self.date)._get_to_invoice_qty()
+                    o.order_line.with_context(accrual_entry_date=self.date)._compute_qty_to_invoice()
                 lines = o.order_line.filtered(
                     lambda l: fields.Float.compare(
                         l.qty_to_invoice,