--- a/addons/event_booth_sale/models/sale_order_line.py
+++ b/addons/event_booth_sale/models/sale_order_line.py
@@ -64,10 +64,9 @@ class SaleOrderLine(models.Model):
         if self.event_booth_pending_ids and (not self.event_id or self.event_id != self.event_booth_pending_ids.event_id):
             self.event_booth_pending_ids = None
 
-    @api.onchange('event_booth_pending_ids')
-    def _onchange_event_booth_pending_ids(self):
-        # trigger the description computation
-        self.product_id_change()
+    @api.depends('event_booth_pending_ids')
+    def _compute_name(self):
+        super()._compute_name()
 
     def _update_event_booths(self, set_paid=False):
         for so_line in self.filtered('is_event_booth'):