--- a/addons/l10n_ar/models/account_fiscal_position.py
+++ b/addons/l10n_ar/models/account_fiscal_position.py
@@ -12,30 +12,26 @@ class AccountFiscalPosition(models.Model):
         'should be auto-detected')
 
     @api.model
-    def get_fiscal_position(self, partner_id, delivery_id=None):
+    def _get_fiscal_position(self, partner, delivery=None):
         """ Take into account the partner afip responsibility in order to auto-detect the fiscal position """
-        company = self.env.company
-        if company.country_id.code == "AR":
-            PartnerObj = self.env['res.partner']
-            partner = PartnerObj.browse(partner_id)
+        if self.env.company.country_id.code != "AR":
+            return super()._get_fiscal_position(partner, delivery=delivery)
 
-            # if no delivery use invoicing
-            if delivery_id:
-                delivery = PartnerObj.browse(delivery_id)
-            else:
-                delivery = partner
+        # if no delivery use invoicing
+        if not delivery:
+            delivery = partner
 
-            # partner manually set fiscal position always win
-            if delivery.property_account_position_id or partner.property_account_position_id:
-                return delivery.property_account_position_id or partner.property_account_position_id
-            domain = [
-                ('auto_apply', '=', True),
-                ('l10n_ar_afip_responsibility_type_ids', '=', self.env['res.partner'].browse(
-                    partner_id).l10n_ar_afip_responsibility_type_id.id),
-                ('company_id', '=', company.id),
-            ]
-            return self.sudo().search(domain, limit=1)
-        return super().get_fiscal_position(partner_id, delivery_id=delivery_id)
+        # partner manually set fiscal position always win
+        manual_fiscal_position = delivery.property_account_position_id or partner.property_account_position_id
+        if manual_fiscal_position:
+            return manual_fiscal_position
+
+        domain = [
+            ('auto_apply', '=', True),
+            ('l10n_ar_afip_responsibility_type_ids', '=', partner.l10n_ar_afip_responsibility_type_id.id),
+            ('company_id', '=', self.env.company.id),
+        ]
+        return self.sudo().search(domain, limit=1)
 
     @api.onchange('l10n_ar_afip_responsibility_type_ids', 'country_group_id', 'country_id', 'zip_from', 'zip_to')
     def _onchange_afip_responsibility(self):