… addons/account/tests/test_fiscal_position.py
… addons/account/tests/test_settings.py
… addons/event_sale/tests/test_event_sale.py
… addons/purchase_stock/models/stock_rule.py
… addons/repair/models/repair.py
… addons/sale/models/account_move_line.py
… addons/sale/models/product_template.py
… addons/sale/models/sale_order.py
… addons/sale/models/sale_order_line.py
… addons/sale/populate/sale_order.py
… addons/sale/tests/__init__.py
… addons/sale/tests/test_onchange.py
… addons/sale/tests/test_performance.py
… addons/sale/tests/test_reinvoice.py
… addons/sale/tests/test_sale_flow.py
… addons/sale/tests/test_sale_order.py
… addons/sale/tests/test_sale_pricelist.py
… addons/sale/tests/test_sale_refund.py
… addons/sale/tests/test_sale_to_invoice.py
… addons/sale/views/sale_views.xml
… addons/sale/wizard/sale_make_invoice_advance.py
… addons/sale_coupon/models/sale_order.py
… addons/sale_coupon/tests/test_program_numbers.py
… addons/sale_expense/tests/test_sale_expense.py
… addons/sale_gift_card/models/sale_order.py
… addons/sale_management/models/sale_order.py
… addons/sale_margin/models/sale_order.py
… addons/sale_product_configurator/tests/test_sale_product_configurator_ui.py
… addons/sale_product_matrix/models/sale_order.py
… addons/sale_project/models/project.py