@robodoo 2021-12-01
Merge method set to merge directly, using the PR as merge commit message